Improved subscription to local/global scene
To prepare to code a (decent) solution for the commits, now the Local and Global creation of the object makes it subscribe to the correct scene: when an obj is created in the local scene, it subscribes to the local events. The same applies to the global scene (it still in WIP, but the foundation is there). 
Before these changes, each obj subscribed on Awake(), but the solution was not good since in the future I want to spawn an object in the global scene sent in the network, and the object itself is not aware of the current scene! So this was the best solution