Commit notification
Show a dialog to the user in case a commit is received. To display the dialog, the notification button is activated (to improve - make it just desaturated) and the user clicks on the button. A dialog with 2 options is displayed: one button takes the user to the global scene to see changes (show me), the other one just closes the dialog (dismiss)