Obj menus in local and global scene
Objects now have two menus: one that is displayed in the local scene and one in the global scene. 
It was important to NOT disable the ObjectManipolator script in the global scene, because disabling it would have made impossible to exploit the OnManipulationStarted event, thus to display the menu in the global scene
Changing scenes also unactivates the visible menu.