Color change on prefab spawn works!
When creating an object in the local layer the user can change its color at runtime! I refactored all the object prewab spawn: now is dynamic and can populate the menus automatically when requested. 
To fix: position of the color menu
To do: make the user change color of an existing object (yikes); try everything on network