Changed global deletion
In case of deletion of only a global object, the obj was copied in the local scene of the user that issued the delete. Now the obj is copied in the local scene of the other user, not the one that issued the delete. This is because if I issue a delete, I want to delete the obj for sure, but maybe the other user doesn't want to erase it completely.