WIP
bugs: sometimes, the color of the object does not show up on decline commit (especially for who issued the commit). Also: duplicating the obj, deleting the original and commiting the new one, makes it disappear from the local scene??? what. Also: declining a commit sometimes make it disappear from the global scene - throws a key not found in globallistmementos