Fixed deletion bug
fixed the deletion bug that would unsubscribe all the object from that scene. Fixed by not unsubscribing (aka: no removeListener()) (lol), and bot destroying the obj, but just hide it from the user (ugly but it works, don't have time for other solutions tbh. A good solution may be resubscribing all the object to that scene when an unbuscribe is prompted, by looping in the list of objects subscribed