Tried using MR Lighting tool (works pretty well) + gaze usage (more on desc)
Lighting tool works ok, sometimes lags but it can be mitigated by taking less pictures? Also, the user gaze works: I can record which obj is targeted and also I can use the gaze direction to make the object spawn in front of the user! And that works pretty well